fixes following code review
